Update docker-compose.yml
Current file doesn't have a proper YAML document structure, it only has a single property (dailytxt) which is not correct.  The entire contents needs to be wrapped up in a top-level key to create a valid YAML document.  This should now be a valid YAML document and can be used in a docker-compose.yml file accross Windows, Linux, Mac machines ⚡ :godmode: ⚡